fix test with authorization
