+++ config/packages/security.yaml
@@ -41,8 +41,8 @@ security:
         # Easy way to control access for large sections of your site
         # Note: Only the *first* access control that matches will be used
     access_control:
-        - { path: ^/api/admin, roles: PUBLIC_ACCESS }
-        - { path: ^/api/users, roles: PUBLIC_ACCESS }
+        - { path: ^/api/admin, roles: ROLE_ADMIN }
+        - { path: ^/api/users, roles: ROLE_USER }
         - { path: ^/api, roles: PUBLIC_ACCESS }
 
     # https://symfony.com/doc/current/security.html#registering-the-user-hashing-passwords
