+++ src/Controller/Api/Admin/ProductController.php
@@ -17,7 +17,11 @@ use JMS\Serializer\SerializationContext;
 use JMS\Serializer\SerializerBuilder;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 
+/**
+ * @IsGranted("ROLE_ADMIN")
+ */
 class ProductController extends AbstractFOSRestController
 {
     public const PRODUCT_PER_PAGE = 10;
