+++ src/Controller/Api/HomePageController.php
@@ -12,7 +12,11 @@ use JMS\Serializer\SerializationContext;
 use JMS\Serializer\SerializerBuilder;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 
+/**
+ * @IsGranted("ROLE_USER")
+ */
 class HomePageController extends AbstractFOSRestController
 {
     public const PRODUCT_PER_PAGE = 9;
